// condoplanner-app/src/apiClient/models/ResidentialLinkDto.ts
export interface ResidentialLinkDto {
    /**
     * 
     * @type {number}
     * @memberof ResidentialLinkDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ResidentialLinkDto
     */
    usuarioId?: number;
    /**
     * 
     * @type {string}
     * @memberof ResidentialLinkDto
     */
    userName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ResidentialLinkDto
     */
    condominiumId?: number;
    /**
     * 
     * @type {string}
     * @memberof ResidentialLinkDto
     */
    condominiumName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ResidentialLinkDto
     */
    apartmentId?: number;
    /**
     * 
     * @type {string}
     * @memberof ResidentialLinkDto
     */
    apartmentNumber?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ResidentialLinkDto
     */
    blockId?: number;
    /**
     * 
     * @type {string}
     * @memberof ResidentialLinkDto
     */
    blockName?: string | null;
    /**
     * 
     * @type {ETipoOcupacao}
     * @memberof ResidentialLinkDto
     */
    occupationType?: ETipoOcupacao;
    /**
     * 
     * @type {boolean}
     * @memberof ResidentialLinkDto
     */
    active?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ResidentialLinkDto
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ResidentialLinkDto
     */
    endDate?: Date | null;
    /**
     * 
     * @type {EStatusVinculoResidencial}
     * @memberof ResidentialLinkDto
     */
    status?: EStatusVinculoResidencial;
}