// condoplanner-app/src/apiClient/models/ETipoOcupacao.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ETipoOcupacao = {
    Proprietario: 'PROPRIETARIO',
    Inquilino: 'INQUILINO',
    Familiar: 'FAMILIAR'
} as const;
export type ETipoOcupacao = typeof ETipoOcupacao[keyof typeof ETipoOcupacao];


export function instanceOfETipoOcupacao(value: any): boolean {
    for (const key in ETipoOcupacao) {
        if (Object.prototype.hasOwnProperty.call(ETipoOcupacao, key)) {
            if (ETipoOcupacao[key as keyof typeof ETipoOcupacao] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ETipoOcupacaoFromJSON(json: any): ETipoOcupacao {
    return ETipoOcupacaoFromJSONTyped(json, false);
}

export function ETipoOcupacaoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ETipoOcupacao {
    return json as ETipoOcupacao;
}

export function ETipoOcupacaoToJSON(value?: ETipoOcupacao | null): any {
    return value as any;
}

export function ETipoOcupacaoToJSONTyped(value: any, ignoreDiscriminator: boolean): ETipoOcupacao {
    return value as ETipoOcupacao;
}